Test#59/feature backend tests

+++ backend/memetory/src/main/java/com/example/memetory/domain/memes/dto/MemesServiceDto.java
@@ -15,16 +15,16 @@ public class MemesServiceDto {
     private Long memeId;
     private String email;
     private String title;
-    private int likeCount;
-    private int commentCount;
+    private Long likeCount;
+    private Long commentCount;
 
     public Memes toEntity(Member member, Meme meme) {
         return Memes.builder()
                 .member(member)
                 .meme(meme)
                 .title(this.title)
-                .likeCount(0)
-                .commentCount(0)
+                .likeCount(0L)
+                .commentCount(0L)
                 .build();
     }
 }
